requirements.txt
Pillow==5.3.0
PySDL2==0.9.6
